requirements.txt
Django>=1.11.7
djangorestframework>=3.7.3
psycopg2>=2.7.3.2
drf-extensions>=0.3.1
